requirements.txt
reflex==0.5.10
openai
python-decouple